updated shopping-list
