+++ src/partials/shopping-list-in-work.html
@@ -1,7 +1,7 @@
 <div class="container-cards">
     <section class="shopping-list">
 
-        <div class="common ">
+        <div class="common">
             <h1 class="main-title reset">Shopping
                 <span class="add-list-color">List</span>
             </h1>
